backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
chromadb==0.4.18
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.2
huggingface-hub==0.15.1
safetensors==0.3.1
numpy==1.24.3
anthropic==0.25.9
httpx>=0.24.0,<0.26.0
PyPDF2==3.0.1
python-docx==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
